+++ src/Common/PoolWithFailoverBase.h
@@ -343,7 +343,7 @@ struct PoolWithFailoverBase<TNestedPool>::PoolState
     Int64 config_priority = 1;
     /// Priority from the GetPriorityFunc.
     Int64 priority = 0;
-    UInt32 random = 0;
+    UInt64 random = 0;
 
     void randomize()
     {
@@ -357,7 +357,7 @@ struct PoolWithFailoverBase<TNestedPool>::PoolState
     }
 
 private:
-    std::minstd_rand rng = std::minstd_rand(randomSeed());
+    std::minstd_rand rng = std::minstd_rand(static_cast<uint_fast32_t>(randomSeed()));
 };
 
 template <typename TNestedPool>
