+++ src/Disks/IO/AsynchronousReadIndirectBufferFromRemoteFS.h
@@ -62,7 +62,7 @@ private:
 
     AsynchronousReaderPtr reader;
 
-    Int32 priority;
+    Int64 priority;
 
     std::shared_ptr<ReadBufferFromRemoteFSGather> impl;
