+++ src/Processors/Formats/Impl/AvroRowOutputFormat.cpp
@@ -203,7 +203,7 @@ AvroSerializer::SchemaWithSerializeFn AvroSerializer::createSchemaWithSerializeF
         case TypeIndex::FixedString:
         {
             auto size = data_type->getSizeOfValueInMemory();
-            auto schema = avro::FixedSchema(size, "fixed_" + toString(type_name_increment));
+            auto schema = avro::FixedSchema(static_cast<int>(size), "fixed_" + toString(type_name_increment));
             return {schema, [](const IColumn & column, size_t row_num, avro::Encoder & encoder)
             {
                 StringRef s = assert_cast<const ColumnFixedString &>(column).getDataAt(row_num);
