+++ src/Server/TCPHandler.h
@@ -148,7 +148,7 @@ private:
     UInt64 client_version_major = 0;
     UInt64 client_version_minor = 0;
     UInt64 client_version_patch = 0;
-    UInt64 client_tcp_protocol_version = 0;
+    UInt32 client_tcp_protocol_version = 0;
 
     /// Connection settings, which are extracted from a context.
     bool send_exception_with_stack_trace = true;
