+++ src/Storages/ExternalStream/Log/FileLog.h
@@ -44,7 +44,7 @@ public:
         ContextPtr context,
         QueryProcessingStage::Enum processed_stage,
         size_t max_block_size,
-        unsigned num_streams) override;
+        size_t num_streams) override;
 
     FileLogSource::FileContainer searchForCandidates();
